+++ quit/web/app.py
@@ -1,11 +1,8 @@
-import sys
 import os
 import urllib
 import hashlib
 import rdflib
 import logging
-import uuid
-import json
 
 from functools import wraps
 
@@ -135,49 +132,14 @@ def register_blueprints(app, config):
     from quit.web.modules.debug import debug
     from quit.web.modules.endpoint import endpoint
     from quit.web.modules.git import git
+    from quit.web.modules.application import application
 
-    for bp in [debug, endpoint, git]:
+    for bp in [debug, endpoint, git, application]:
         app.register_blueprint(bp)
 
     @app.route("/")
     def index():
         return redirect(url_for('git.commits'))
-
-    @app.route("/login")
-    def login():
-        if "state" not in session:
-            state = str(uuid.uuid4())
-            session["state"] = state
-        else:
-            state = session["state"]
-        logger.debug("request url: {}".format(request.url))
-        redirect_uri = request.url
-        authorizeEndpoint = "https://github.com/login/oauth/authorize"
-        tokenEndpoint = "https://github.com/login/oauth/access_token"
-        error = request.values.get('error', None)
-        code = request.values.get('code', None)
-        if error or code is None:
-            params = {'client_id': config.oauthclientid,
-                      'redirect_uri': redirect_uri,
-                      'scope': 'repo',
-                      'state': state}
-            loginURL = authorizeEndpoint + "?" + urllib.parse.urlencode(params)
-            return "<a href='{}'>Login with GitHub</a>".format(loginURL)
-        else:
-            request_state = request.values.get('state', None)
-            if not state == request_state:
-                return "Error"
-            post_data = {'client_id': config.oauthclientid,
-                         'client_secret': config.oauthclientsecret,
-                         'code': code,
-                         'state': state}
-            tokenrequest = urllib.request.Request(tokenEndpoint,
-                                                  urllib.parse.urlencode(post_data).encode())
-            tokenrequest.add_header('Accept', 'application/json')
-            tokenresponse = urllib.request.urlopen(tokenrequest).read().decode("utf-8")
-            token = json.loads(tokenresponse)["access_token"]
-            session["OAUTH_TOKEN"] = token
-            return "success"
 
 
 def register_hook(app):
