+++ quit/web/templates/commits.html
@@ -3,6 +3,13 @@
 {% block headline %}<div class="lead">Commits</div>{% endblock %}
 
 {% block quit_content %}
+<div style="float: right">
+{% if isLoggedIn() %}
+You are Logged in To GitHub (<a href="{{ url_for('application.logout') }}">logout</a>)
+{% else %}
+<a style="float: right" href="{{ url_for('application.login') }}" class="btn btn-default"><i class="fa fa-sign-in" aria-hidden="true"></i> Login with GitHub</a>
+{% endif %}
+</div>
 <form class="form-inline" style="display: inline" id="transmission" role="navigation">
 <ul class="nav nav-tabs">
   <li role="presentation" class="active"><a href="#fetch" action="{{ url_for('.fetch')}}" data-toggle="tab" transmission_method="fetch"><i class="fa fa-cloud-download" aria-hidden="true"></i> Fetch</a></li>
